pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "webrtc-ops-env"
version = "1.0.0"
description = "Simulated WebRTC infrastructure diagnostics environment for AI agent evaluation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "pheelip0030"},
]

dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "fastmcp>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
